+++ ios/Classes/InAppWebView/InAppWebView.swift
@@ -433,7 +433,9 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
             }
             
             if #available(iOS 15.0, *) {
-                configuration.preferences.isSiteSpecificQuirksModeEnabled = settings.isSiteSpecificQuirksModeEnabled
+                if (configuration.preferences.responds(to: #selector(getter: InAppWebViewSettings.isSiteSpecificQuirksModeEnabled))) {
+                    configuration.preferences.isSiteSpecificQuirksModeEnabled = settings.isSiteSpecificQuirksModeEnabled
+                }
             }
         }
     }
@@ -1130,7 +1132,8 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
                let underPageBackgroundColor = newSettings.underPageBackgroundColor, !underPageBackgroundColor.isEmpty {
                 self.underPageBackgroundColor = UIColor(hexString: underPageBackgroundColor)
             }
-            if newSettingsMap["isSiteSpecificQuirksModeEnabled"] != nil &&
+            if configuration.preferences.responds(to: #selector(getter: InAppWebViewSettings.isSiteSpecificQuirksModeEnabled)),
+               newSettingsMap["isSiteSpecificQuirksModeEnabled"] != nil &&
                 settings?.isSiteSpecificQuirksModeEnabled != newSettings.isSiteSpecificQuirksModeEnabled {
                 configuration.preferences.isSiteSpecificQuirksModeEnabled = newSettings.isSiteSpecificQuirksModeEnabled
             }
@@ -1483,6 +1486,83 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
         result["currentIndex"] = currentIndex
         
         return result;
+    }
+
+    @available(iOS 15.0, *)
+    @available(macOS 12.0, *)
+    @available(macCatalyst 15.0, *)
+    public func webView(_ webView: WKWebView,
+                        requestMediaCapturePermissionFor origin: WKSecurityOrigin,
+                        initiatedByFrame frame: WKFrameInfo,
+                        type: WKMediaCaptureType,
+                        decisionHandler: @escaping (WKPermissionDecision) -> Void) {
+        let origin = "\(origin.protocol)://\(origin.host)\(origin.port != 0 ? ":" + String(origin.port) : "")"
+        let permissionRequest = PermissionRequest(origin: origin, resources: [type.rawValue], frame: frame)
+
+        onPermissionRequest(request: permissionRequest, result: {(result) -> Void in
+            if result is FlutterError {
+                print((result as! FlutterError).message ?? "")
+                decisionHandler(.deny)
+            }
+            else if (result as? NSObject) == FlutterMethodNotImplemented {
+                decisionHandler(.deny)
+            }
+            else {
+                var response: [String: Any]
+                if let r = result {
+                    response = r as! [String: Any]
+                    var action = response["action"] as? Int
+                    action = action != nil ? action : 0;
+                    switch action {
+                    case 1:
+                        decisionHandler(.grant)
+                        break
+                    default:
+                        decisionHandler(.deny)
+                    }
+                    return;
+                }
+                decisionHandler(.deny)
+            }
+        })
+    }
+    
+    @available(iOS 15.0, *)
+    @available(macOS 12.0, *)
+    @available(macCatalyst 15.0, *)
+    public func webView(_ webView: WKWebView,
+                        requestDeviceOrientationAndMotionPermissionFor origin: WKSecurityOrigin,
+                        initiatedByFrame frame: WKFrameInfo,
+                        decisionHandler: @escaping (WKPermissionDecision) -> Void) {
+        let origin = "\(origin.protocol)://\(origin.host)\(origin.port != 0 ? ":" + String(origin.port) : "")"
+        let permissionRequest = PermissionRequest(origin: origin, resources: ["deviceOrientationAndMotion"], frame: frame)
+        
+        onPermissionRequest(request: permissionRequest, result: {(result) -> Void in
+            if result is FlutterError {
+                print((result as! FlutterError).message ?? "")
+                decisionHandler(.deny)
+            }
+            else if (result as? NSObject) == FlutterMethodNotImplemented {
+                decisionHandler(.deny)
+            }
+            else {
+                var response: [String: Any]
+                if let r = result {
+                    response = r as! [String: Any]
+                    var action = response["action"] as? Int
+                    action = action != nil ? action : 0;
+                    switch action {
+                    case 1:
+                        decisionHandler(.grant)
+                        break
+                    default:
+                        decisionHandler(.deny)
+                    }
+                    return;
+                }
+                decisionHandler(.deny)
+            }
+        })
     }
     
     @available(iOS 13.0, *)
@@ -2424,6 +2504,10 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate, WKNavi
     
     public func onNavigationResponse(navigationResponse: WKNavigationResponse, result: FlutterResult?) {
         channel?.invokeMethod("onNavigationResponse", arguments: navigationResponse.toMap(), result: result)
+    }
+    
+    public func onPermissionRequest(request: PermissionRequest, result: FlutterResult?) {
+        channel?.invokeMethod("onPermissionRequest", arguments: request.toMap(), result: result)
     }
     
     public func onReceivedHttpAuthRequest(challenge: URLAuthenticationChallenge, result: FlutterResult?) {